Revert stream changes
